LookupJoin reduce inequi copying
